requirements.txt
langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-google-genai>=2.0,<3.0
google-genai>=1.0,<2.0
faiss-cpu>=1.8,<2.0
sentence-transformers>=3.0,<4.0
python-dotenv>=1.0,<2.0
pypdf>=5.0,<6.0
streamlit>=1.40,<2.0
tiktoken>=0.8,<1.0
